Backup just in case
so, i've sorted deal with h2 or bigger header text in form. To make change the font-weight gotta cut out specs like <p> or <h2>, then works properly when adding an div-container and applying styles to it. That helped to center header with other text left-aligned.
NOW i need to force form to close when clicking outside of it (cross button also, but later, it's not that hard now), so messing with positioning, divs and JS code ahead, gotta secure this